+++ crates/compression/src/compressed_block_payload/mod.rs
@@ -1,3 +1,4 @@
 pub mod v0;
 
+#[cfg(feature = "fault-proving")]
 pub mod v1;